fixed ahrs inverse_sqrt computation and tuned

+++ alpha_localization/src/ahrs.cpp
@@ -15,11 +15,12 @@
 
 #include <alpha_localization/ahrs.h>
 #include <math.h>
+#include <stdint.h>
 
 float invSqrt(float x) {
 	float halfx = 0.5f * x;
 	float y = x;
-	long i = *(long*)&y;
+	int32_t i = *(int32_t*)&y;
 	i = 0x5f3759df - (i>>1);
 	y = *(float*)&i;
 	y = y * (1.5f - (halfx * y * y));
